seperate auth path update connected endpoints
