+++ api/__init__.py
@@ -7,6 +7,7 @@ from .routes import router as routes_router
 from .record import router as record_router
 from .member import router as member_router
 from .video import router as video_router
+from .auth import router as auth_router
 # from .test import router as test_router
 
 
@@ -14,4 +15,5 @@ main_router.include_router(routes_router)
 main_router.include_router(record_router)
 main_router.include_router(member_router)
 main_router.include_router(video_router)
+main_router.include_router(auth_router)
 # main_router.include_router(test_router)
